+++ src/Cake.AWS.CodeDeploy/Interfaces/IDeployManager.cs
@@ -1,4 +1,8 @@
-﻿
+﻿#region Using Statements
+using System.Threading;
+using System.Threading.Tasks;
+#endregion
+
 
 
 namespace Cake.AWS.CodeDeploy
@@ -15,7 +19,8 @@ namespace Cake.AWS.CodeDeploy
         /// <param name="applicationName">The name of an AWS CodeDeploy application associated with the applicable IAM user or AWS account.</param>
         /// <param name="deploymentGroup">The name of the deployment group.</param>
         /// <param name="settings">The <see cref="DeploySettings"/> used during the request to AWS.</param>
-        bool CreateDeployment(string applicationName, string deploymentGroup, DeploySettings settings);
+        /// <param name="cancellationToken">A cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
+        Task<bool> CreateDeployment(string applicationName, string deploymentGroup, DeploySettings settings, CancellationToken cancellationToken = default(CancellationToken));
 
 
 
@@ -24,7 +29,8 @@ namespace Cake.AWS.CodeDeploy
         /// </summary>
         /// <param name="applicationName">The name of an AWS CodeDeploy application associated with the applicable IAM user or AWS account.</param>
         /// <param name="settings">The <see cref="DeploySettings"/> used during the request to AWS.</param>
-        bool RegisterApplicationRevision(string applicationName, DeploySettings settings);
+        /// <param name="cancellationToken">A cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
+        Task<bool> RegisterApplicationRevision(string applicationName, DeploySettings settings, CancellationToken cancellationToken = default(CancellationToken));
         #endregion
     }
 }